fix: + custom link after merge error
