+++ src/app/[locale]/page.tsx
@@ -1,27 +1,12 @@
 'use client';
-
-import { ReactNode, FC } from 'react';
 import { useTranslations } from 'next-intl';
-import { Button, Link } from '@mui/material';
+import { Button } from '@mui/material';
 import Typography from '@mui/material/Typography';
-import { colors } from 'constants/colors';
 import { AppRoute } from 'enums';
 import { typography } from 'theme/typography';
 import WelcomePageContainer from 'components/WelcomePageContainer/WelcomePageContainer';
 import { RightSection, StyledBtnWrapper } from 'components/WelcomePageContainer/styles';
-import { PublicRoute } from 'components/custom';
-
-interface CustomLinkProps {
-  href: string;
-  color?: string;
-  children: ReactNode;
-}
-
-const CustomLink: FC<CustomLinkProps> = ({ href, color = colors.TEST_MAIN_COLOR, children }) => (
-  <Link underline="none" color={color} href={href}>
-    {children}
-  </Link>
-);
+import { CustomLink, PublicRoute } from 'components/custom';
 
 export default function HomePage() {
   const t = useTranslations('Home');
@@ -32,14 +17,10 @@ export default function HomePage() {
         <RightSection maxWidth="sm">
           <StyledBtnWrapper>
             <Button variant="contained" sx={{ marginRight: '10px' }} size="large">
-              <Link underline={'none'} color={colors.TEST_MAIN_COLOR} href={AppRoute.SIGN_IN_PAGE}>
-                {t('Page.signInLink')}
-              </Link>
+              <CustomLink href={AppRoute.SIGN_IN_PAGE}>{t('Page.signInLink')}</CustomLink>
             </Button>
             <Button variant="contained" size="large">
-              <Link underline={'none'} color={colors.TEST_MAIN_COLOR} href={AppRoute.SIGN_UP_PAGE}>
-                {t('Page.signUpLink')}
-              </Link>
+              <CustomLink href={AppRoute.SIGN_UP_PAGE}>{t('Page.signUpLink')}</CustomLink>
             </Button>
           </StyledBtnWrapper>
           <Typography
@@ -55,9 +36,9 @@ export default function HomePage() {
             {t('Page.title')}
           </Typography>
           <Button variant="contained" fullWidth size="large">
-            <Link underline={'none'} color="primary" href={AppRoute.SIGN_IN_PAGE}>
+            <CustomLink color="primary" href={AppRoute.SIGN_IN_PAGE}>
               {t('Page.buttonTxt')}
-            </Link>
+            </CustomLink>
           </Button>
         </RightSection>
       </WelcomePageContainer>
