+++ src/components/custom/index.ts
@@ -1,3 +1,4 @@
 export { App } from './app/app';
 export { PublicRoute } from './public-route/public-route';
 export { ProtectedRoute } from './protected-route/protected-route';
+export { CustomLink } from './custom-link/custom-link';
